Improve: versioning of code dependencies (JUCE).
 - The idea is to have all deps that we build from code part of the repository at all time.

This sounds like a counter-intuitive thing to do, since we should usually try to keep the repo clean of dependencies. But in the context of building plugins and other monolithic apps it makes sense - it's much easier to track issues down to updates of the dependencies, and locate where the issues actually are within the code of those dependencies.

Another benefit is that the code becomes easier to index by tools like ctags for code navigation with vim/emacs/sublimetext.

Having said that, we should (at all cost) avoid making changes to dependecy code. In the context of JUCE, we need to do our best to get whatever is blocking us, fixed within the official JUCE repo and if a deadline prevents us to wait this to be released:
   - we can ask the change to be cherry-picked by JUCE staff to the master branch of the JUCE repo (presumably thus making the full update of JUCE from master safer than taking the bleading-edge dev branch);
   - if the previous option isn't available, or the JUCE team doesn't want to cherry-pick to master - ONLY THEN we would make the change to our copy of JUCE, but not before we create an issue with high severity in our issue tracker (currently the regular Github Issues) to remind us to regularly check for an update so we can get an official (unaltered) release in our repo ASAP.